requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fake-useragent==1.2.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==4.9.3
openpyxl==3.1.2
outcome==1.2.0
Pillow==10.0.0
PySocks==1.7.1
requests==2.31.0
rfc3986==1.5.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
